+++ src/main/java/com/example/apiambienteescolarsql/dto/projection/TurmaMediaProjection.java
@@ -4,5 +4,5 @@ public interface TurmaMediaProjection {
     Long getNCdTurma();
     Integer getISerie();
     String getCNmTurma();
-    Long getNMedia();
+    Double getNMedia();
 }
